+++ src/main/java/com/example/elaborato_ing/AutoUsata.java
@@ -12,7 +12,7 @@ public class AutoUsata extends Auto{
     private int incidenti;
     private Stato interni;
 
-    public AutoUsata(Marca marca, Modello modello, double lunghezza, double altezza, double larghezza, double peso, double volumeBagagliaio, Motore motore, double costo, String sconto) {
+    public AutoUsata(Marca marca, Modello modello, double lunghezza, double altezza, double larghezza, double peso, double volumeBagagliaio, Motore motore, int costo, String sconto) {
         super(marca, modello, lunghezza, altezza, larghezza, peso, volumeBagagliaio, motore, costo, sconto);
     }
 }
